gtexslte: fixing AOSP kernel build & ext WiFi mod
This requires the updated build_kernel.sh version as well to
function properly.
Building an external module like WiFi requires a full kernel
build before (make modules etc is *not* enough) and so it must
either happen within the kernel target or ramdisk. I decided to
put it in the ramdisk target which happens after the kernel build.
In order to make it fully work though Module.symvers is needed which
gets created by a full kernel build.
Even when you do not do a full kernel build the WiFi module can be
created but you will see errors like:

> insmod: failed to load /lib/modules/sprdwl.ko: Exec format error

when trying to load it manually (insmod /lib/modules/sprdwl.ko).
This is due to the fact the module was build without a full/correct
Module.symvers.

That's why customizing the ramdisk target is done in mkbootimg.mk
which runs after kernel build, copies the Module.symvers, building
sprdwl.ko and finally copying it over to the ramdisk.

I guess .. very likely.. there are better ways. the current
implementation works and that's all I need for the moment.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>