gtexslte: placing sprdwl module in /system
switched to /system for the WiFi module. That module is tricky to get build within the
regular AOSP build even with the modifications in the kernel tree.

this + the fact that it is not needed to have these in the RAMDISK makes my decision
to put all modules to /system instead.

for WiFi:

 * /lib/modules: must run *before* INSTALLED_RAMDISK_TARGET and *after* TARGET_KERNEL_BINARIES
 * /system/lib/modules: simply after TARGET_KERNEL_BINARIES

There might be a way without modding build/ and/or vendor/ in the source tree but I haven't
found any and this gets the job done.

Signed-off-by: steadfasterX <steadfasterX@gmail.com>